Organize classes into namespaces Create Autofac modules
