--- SQLTableClassGenerator/ITableClassBuilder.cs
@@ -1,7 +1,0 @@
-﻿namespace SQLTableClassGenerator
-{
-    public interface ITableClassBuilder
-    {
-        string Build(TableDef table);
-    }
-}
