--- SQLTableClassGenerator/ITableClassPart.cs
@@ -1,7 +1,0 @@
-﻿namespace SQLTableClassGenerator
-{
-    public interface ITableClassPart
-    {
-        string GetString(TableDef table);
-    }
-}
