+++ SQLTableClassGenerator/TableClassParts/ITableClassBuilder.cs
@@ -1,0 +1,9 @@
+﻿using SQLTableClassGenerator.TableElements;
+
+namespace SQLTableClassGenerator.TableClassParts
+{
+    public interface ITableClassBuilder
+    {
+        string Build(TableDef table);
+    }
+}
