+++ SQLTableClassGenerator/TableClassParts/ITableClassPart.cs
@@ -1,0 +1,9 @@
+﻿using SQLTableClassGenerator.TableElements;
+
+namespace SQLTableClassGenerator.TableClassParts
+{
+    public interface ITableClassPart
+    {
+        string GetString(TableDef table);
+    }
+}
