+++ SQLTableClassGenerator/UI/MainForm.resx
