+++ SQLTableClassGenerator/UI/UIModule.cs
@@ -1,0 +1,14 @@
+﻿using System.Windows.Forms;
+using Autofac;
+
+namespace SQLTableClassGenerator.UI
+{
+    internal class UIModule : Module
+    {
+        protected override void Load(ContainerBuilder builder)
+        {
+            builder.RegisterType<TreeViewPopulator>().AsImplementedInterfaces();
+            builder.RegisterType<MainForm>().As<Form>();
+        }
+    }
+}
